+++ Changes.md
@@ -6,6 +6,10 @@ Improvements
 
 - TabbedContainer : Added menu button to allow selection of tabs that are not
   visible due to a lack of horizontal space.
+- ImageView : Added multiple color inspectors.  Ctrl-click on an image to create
+  a pixel inspector, or Ctrl-drag to create an area inspector.  The image Viewer
+  footer now shows the results from all your inspectors, and allows you to add or
+  delete them.
 
 Fixes
 -----